refactor: fixed naming BE and FE
